code_resources: also prioritize rules based on order
One of the first rules with weight 20 simply matches '^' (i.e.
everything) so without removing that rule (which we don't want,
considering that we see Apple's codesign tool emits the same rule) it
implies we should consider later rules to have a higher priority if they
have the same weight.

With this change to the interpretation of rule weights we now omit
Info.plist and PkgInfo from the rules2 section consistent with the
codesign tool.